requirements.txt
amqp==5.3.1
billiard==4.2.4
blinker==1.9.0
Celery==5.6.2 # Autorise less der45td	versions récentes jusqu'à la v7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.5.3
Werkzeug==3.1.5
yt-dlp==2026.1.31
